requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
ffmpeg-python
pydub
pytest

